requirements.txt
requests
beautifulsoup
google-generativeai
python-dotenv
pandas
openai
